deleted old player & fixed shared
I still have to create some incognito mode for the player, to only display like buttons if the user is connected
