+++ hooks/player/use-player.ts
@@ -1,5 +1,4 @@
 import { create } from "zustand";
-import { Wave as WaveObject } from "@foobar404/wave";
 
 type playHandler = () => void;
 
@@ -23,10 +22,6 @@ interface PlayerStore {
     // Other values
     volume: number;
     setVolume: (volume: number) => void;
-
-    wave: WaveObject | null;
-    canvas: HTMLCanvasElement | null;
-    setWave: (wave: WaveObject, canvas: HTMLCanvasElement) => void;
 
     // Handlers
     handlePlay: playHandler;
@@ -53,12 +48,6 @@ const usePlayer = create<PlayerStore>((set, getState) => ({
     // Other values
     volume: 1,
     setVolume: (volume: number) => set({ volume: volume }),
-    wave: null,
-    canvas: null,
-    setWave: (wave: WaveObject, canvas: HTMLCanvasElement) => {
-        set({ wave: wave });
-        set({ canvas: canvas });
-    },
 
     // Handlers
     handlePlay: () => { console.log('init handlePlay called') },
