+++ lib/components/player/Player.tsx
@@ -1,28 +1,67 @@
 'use client';
 
-import usePlayer from "@/hooks/player/use-player";
-import useGetSongById from "@/hooks/use-get-song-by-id";
-import PlayerContentWrapper from "@/lib/components/player/PlayerContentWrapper";
+import React, {useCallback, useEffect, useState} from 'react';
 
-const Player = () => {
-    const player = usePlayer();
-    const { song } = useGetSongById(player.activeId);
+import usePlayer from '@/hooks/player/use-player';
+import useGetSongById from '@/hooks/use-get-song-by-id';
+import { StandardPlayer } from '@/lib/components/player/StandardPlayer';
+import { BigPicturePlayer } from '@/lib/components/player/BigPicturePlayer';
 
-    if (!song || !player.activeId) {
-        return null;
+export const Player = (): React.JSX.Element | null => {
+  const player = usePlayer();
+  const { song  } = useGetSongById(player.activeId);
+  const [ audioPlayer, setAudioPlayer ] = useState<HTMLAudioElement | null>(null);
+
+  const onEndReached = useCallback(() => {
+    if (player.onRepeat) return;
+
+    const currentIndex = player.ids.findIndex((id) => id === player.activeId);
+    const nextSong = player.ids[currentIndex + 1];
+
+    if (!nextSong) {
+      // Go back to start of the queue
+      player.setId(player.ids[0]);
+    } else {
+      player.setId(nextSong);
     }
+  }, [player]);
+  
+  useEffect(() => {
+    if (song && !audioPlayer) {
+      const audio = new Audio(song.songUrl);
 
-    return (
-        <div
-            className={`fixed bottom-0 bg-black w-full py-2 h-[80px] px-4 transition-all ${ player.bigPicture && 'h-full'} md:h-[80px]`}
-        >
-            {/* PlayerContent has a key value as if the key changes, it re-renders */}
-            <PlayerContentWrapper
-                key={song.songUrl}
-                song={song}
-            />
-        </div>
-    );
+      audio.onended = onEndReached;
+      audio.textContent = song.songUrl;
+
+      setAudioPlayer(audio);
+
+      // When a music is loaded, start playing it directly
+      audio.play().then((value) => player.setIsPlaying(true));
+    } else if (song && audioPlayer && audioPlayer.textContent !== song.songUrl) {
+      // Then this player is not fresh, the user switched the song manually
+
+      audioPlayer.src = song.songUrl;
+      audioPlayer.onended = onEndReached; // We have to update the audioPlayer onend function with up-to-date info
+      audioPlayer.textContent = song.songUrl;
+      audioPlayer.play().then((value) => player.setIsPlaying(true));
+    }
+  }, [song, audioPlayer, player, onEndReached]);
+
+  if (!song || !player.activeId) {
+    return null;
+  }
+
+  return (
+    <div
+      className={`fixed bottom-0 bg-black w-full py-2 h-[80px] px-4 transition-all flex items-center md:justify-between ${player.bigPicture && 'h-full'}`}
+      key={song.songUrl}
+    >
+      {audioPlayer && (
+        <>
+          <StandardPlayer audioPlayer={audioPlayer} song={song} />
+          <BigPicturePlayer audioPlayer={audioPlayer} song={song} />
+        </>
+      )}
+    </div>
+  );
 }
-
-export default Player;
