+++ lib/components/player/StandardPlayer.tsx
@@ -3,8 +3,8 @@ import MediaItem from '@/lib/components/MediaItem';
 import LikeButton from '@/lib/components/LikeButton';
 import { Song } from '@prisma/client';
 import { Pause, Play, Repeat2, Shuffle, SkipBack, SkipForward, Maximize2 } from 'lucide-react';
-import { ProgressBar } from '@/lib/components/new-player/ProgressBar';
-import { VolumeBar } from '@/lib/components/new-player/VolumeBar';
+import { ProgressBar } from '@/lib/components/player/ProgressBar';
+import { VolumeBar } from '@/lib/components/player/VolumeBar';
 import { usePlayerControls } from '@/hooks/player/use-player-new-controls';
 import usePlayer from '@/hooks/player/use-player';
 import { getBreakpointValue } from '@/lib/getBreakpointValue';
