+++ lib/components/player/VolumeBar.tsx
